requirements.txt
python-dotenv
PyPDF2
langchain
pinecone

langchain-groq
sentence-transformers
